requirements.txt
flask
moviepy
speechrecognition
openai
python-docx
